[FIX] account/saas~17.5: set correct column type
Update `the amount_untaxed_in_currency_signed` field
to the numeric type in the account_move table to address
overflow issues encountered during the upgrade process.
This change resolves the "integer out of range" error.
According to standard practices, monetary fields should
be of type numeric. The error occurred in customer
databases where unrelairet real numbers were being used.

[ref](https://github.com/odoo/odoo/blob/18.0/odoo/fields.py#L1652-L1654)
[field](https://github.com/odoo/odoo/blob/18.0/addons/account/models/account_move.py#L497)

```
Traceback (most recent call last):
  File "/home/odoo/src/odoo/18.0/odoo/service/server.py", line 1305, in preload_registries
    registry = Registry.new(dbname, update_module=update_module)
  File "<decorator-gen-13>", line 2, in new
  File "/home/odoo/src/odoo/18.0/odoo/tools/func.py", line 97, in locked
    return func(inst, *args, **kwargs)
  File "/home/odoo/src/odoo/18.0/odoo/modules/registry.py", line 131, in new
    odoo.modules.load_modules(registry, force_demo, status, update_module)
  File "/home/odoo/src/odoo/18.0/odoo/modules/loading.py", line 475, in load_modules
    processed_modules += load_marked_modules(env, graph,
  File "/home/odoo/src/odoo/18.0/odoo/modules/loading.py", line 363, in load_marked_modules
    loaded, processed = load_module_graph(
  File "/home/odoo/src/odoo/18.0/odoo/modules/loading.py", line 180, in load_module_graph
    migrations.migrate_module(package, 'pre')
  File "/home/odoo/src/odoo/18.0/odoo/modules/migration.py", line 254, in migrate_module
    mod.migrate(self.cr, installed_version)
  File "/tmp/tmpwbn08iy2/migrations/account/saas~17.5.1.2/pre-migrate.py", line 117, in migrate
    util.explode_execute(cr, query_amount_untaxed, table="account_move")
  File "/tmp/tmpwbn08iy2/migrations/util/pg.py", line 350, in explode_execute
    return parallel_execute(
  File "/tmp/tmpwbn08iy2/migrations/util/pg.py", line 175, in parallel_execute
    return parallel_execute_impl(cr, queries, logger=_logger)
  File "/tmp/tmpwbn08iy2/migrations/util/pg.py", line 92, in _parallel_execute_threaded
    cr.execute(queries[0])
  File "/home/odoo/src/odoo/18.0/odoo/sql_db.py", line 354, in execute
    res = self._obj.execute(query, params)
psycopg2.errors.NumericValueOutOfRange: integer out of range

bda@(none):bda_2161609> SELECT amount_untaxed
 FROM account_move
 WHERE amount_untaxed > 2147483647 OR amount_untaxed < -2147483648;
+----------------+
| amount_untaxed |
|----------------|
| 5424432432.43  |
+----------------+
SELECT 1
Time: 0.009s

upg-2161609
tbg-1586
```

closes odoo/upgrade#6668

+++ migrations/account/saas~17.5.1.2/pre-migrate.py
@@ -107,7 +107,7 @@ def migrate(cr, version):
     util.invert_boolean_field(cr, "account.move", "to_check", "checked")
     cr.execute("ALTER INDEX IF EXISTS account_move_to_check_idx RENAME TO account_move_checked_idx")
 
-    util.create_column(cr, "account_move", "amount_untaxed_in_currency_signed", "int4")
+    util.create_column(cr, "account_move", "amount_untaxed_in_currency_signed", "numeric")
     query_amount_untaxed = """
         UPDATE account_move
            SET amount_untaxed_in_currency_signed = amount_untaxed *
